Refactor and update version before change
