+++ pom.xml
@@ -192,7 +192,7 @@
             <dependency>
                 <groupId>au.org.aodn</groupId>
                 <artifactId>stacmodel</artifactId>
-                <version>0.0.60</version>
+                <version>0.0.63</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
